+++ src/helpers/constants.py
@@ -1340,5 +1340,6 @@ class ScheduledJobType(str, Enum):
     SYSTEM = 'system'
 
 
+VERSION_NORM_LENGTH = 6
 TENANTS_QUERY_THRESHOLD = 20
-
+LATEST_VERSION_TAG = ':'  # ':' > '999999.999999.999999'