+++ src/services/rule_meta_service.py
@@ -2,13 +2,25 @@ from datetime import datetime
 from typing import Optional, Iterator, Generator, Iterable, Any, Literal
 
 from pydantic import BaseModel, Field, ConfigDict
-from modular_sdk.models.pynamongo.adapter import ResultIterator, EmptyResultIterator
+from modular_sdk.models.pynamongo.adapter import (
+    ResultIterator,
+    EmptyResultIterator,
+)
 from pynamodb.expressions.condition import Condition
 
 from helpers import adjust_cloud
-from helpers.constants import COMPOUND_KEYS_SEPARATOR, ID_ATTR, NAME_ATTR, \
-    VERSION_ATTR, FILTERS_ATTR, LOCATION_ATTR, CLOUD_ATTR, COMMENT_ATTR, \
-    RuleDomain
+from helpers.constants import (
+    COMPOUND_KEYS_SEPARATOR,
+    ID_ATTR,
+    NAME_ATTR,
+    VERSION_ATTR,
+    FILTERS_ATTR,
+    LOCATION_ATTR,
+    CLOUD_ATTR,
+    COMMENT_ATTR,
+    RuleDomain,
+    LATEST_VERSION_TAG,
+)
 from helpers.log_helper import get_logger
 from helpers.time_helper import utc_iso
 from models.rule import Rule
@@ -48,6 +60,7 @@ class RuleName:
     """
     Represents rule name scheme used by security team.
     """
+
     known_clouds = {'aws', 'azure', 'gcp', 'k8s'}  # inside rule name
     Resolved = tuple[str | None, str | None, str | None, str | None]
     __slots__ = ('_raw', '_resolved')
@@ -131,9 +144,12 @@ class RuleNamesResolver:  # TODO test
     __slots__ = ('_available_ids', '_allow_multiple', '_allow_ambiguous')
     Payload = tuple[str, bool]
 
-    def __init__(self, resolve_from: Iterable[str],
-                 allow_multiple: bool = False,
-                 allow_ambiguous: bool = False):
+    def __init__(
+        self,
+        resolve_from: Iterable[str],
+        allow_multiple: bool = False,
+        allow_ambiguous: bool = False,
+    ):
         """
         :param allow_multiple: whether to allow to resolve multiple rules
         from one provided name (in case the name is ambiguous)
@@ -142,8 +158,10 @@ class RuleNamesResolver:  # TODO test
         :param resolve_from: Iterable of rules to resolve from
         """
         if allow_ambiguous and allow_multiple:
-            raise AssertionError('If allow_multiple is True, '
-                                 'allow_ambiguous must not be provided')
+            raise AssertionError(
+                'If allow_multiple is True, '
+                'allow_ambiguous must not be provided'
+            )
         self._available_ids = resolve_from
         self._allow_multiple = allow_multiple
         self._allow_ambiguous = allow_ambiguous
@@ -167,8 +185,10 @@ class RuleNamesResolver:  # TODO test
             return
         # resolved rules exist
         if self._allow_multiple:
-            _LOG.debug(f'Multiple rules from one name are allowed. '
-                       f'Resolving all from {name}')
+            _LOG.debug(
+                f'Multiple rules from one name are allowed. '
+                f'Resolving all from {name}'
+            )
             for rule in resolved:
                 yield rule, True
             return
@@ -182,8 +202,9 @@ class RuleNamesResolver:  # TODO test
         _LOG.warning(f'Cannot certainly resolve name: {name}')
         yield name, False
 
-    def resolve_multiple_names(self, names: Iterable[str]
-                               ) -> Generator[Payload, None, None]:
+    def resolve_multiple_names(
+        self, names: Iterable[str]
+    ) -> Generator[Payload, None, None]:
         """
         Yields tuples there the first element is either resolved or not
         resolved rule name. If it's not resolved, the same value as came
@@ -202,25 +223,27 @@ class RuleNamesResolver:  # TODO test
         for name in names:
             yield from self.resolve_one_name(name)
 
-    def resolved_names(self, names: Iterable[str]
-                       ) -> Generator[str, None, None]:
+    def resolved_names(
+        self, names: Iterable[str]
+    ) -> Generator[str, None, None]:
         """
         Ignores whether the rule was resolved or not. Just tries to do it
         :param names:
         :return:
         """
-        yield from (
-            name for name, _ in self.resolve_multiple_names(names)
-        )
+        yield from (name for name, _ in self.resolve_multiple_names(names))
 
 
 class RuleService(BaseDataService[Rule]):
     FilterValue = str | set[str] | tuple[str]
 
     @staticmethod
-    def gen_rule_id(customer: str, cloud: Optional[str] = None,
-                    name: Optional[str] = None,
-                    version: Optional[str] = None) -> str:
+    def gen_rule_id(
+        customer: str,
+        cloud: Optional[str] = None,
+        name: Optional[str] = None,
+        version: Optional[str] = None,
+    ) -> str:
         """
         Make sure to supply this method with ALL the attributes in case
         you create a new rule
@@ -231,8 +254,10 @@ class RuleService(BaseDataService[Rule]):
         :return:
         """
         if name and not cloud:
-            _LOG.warning('Cloud was not provided but name was. '
-                         'Trying to resolve cloud from name')
+            _LOG.warning(
+                'Cloud was not provided but name was. '
+                'Trying to resolve cloud from name'
+            )
             cloud = RuleName(name).cloud.value
         if name and not cloud or version and not name:
             raise AssertionError('Invalid usage')
@@ -242,14 +267,16 @@ class RuleService(BaseDataService[Rule]):
             _id += f'{cloud}{COMPOUND_KEYS_SEPARATOR}'
         if name:
             _id += f'{name}{COMPOUND_KEYS_SEPARATOR}'
-        if version:  # or Rule.latest_version(), or any str, or None
+        if version:
             _id += f'{version}'
         return _id
 
     @staticmethod
-    def gen_location(git_project: Optional[str] = None,
-                     ref: Optional[str] = None,
-                     path: Optional[str] = None) -> str:
+    def gen_location(
+        git_project: Optional[str] = None,
+        ref: Optional[str] = None,
+        path: Optional[str] = None,
+    ) -> str:
         """
         Make sure to supply this method with ALL the attribute in case
         you create a new rule
@@ -269,22 +296,26 @@ class RuleService(BaseDataService[Rule]):
             loc += f'{path}'
         return loc
 
-    def create(self, customer: str, name: str, resource: str, description: str,
-               rule_source_id: str,
-               cloud: Optional[str] = None,
-               filters: Optional[list[dict]] = None,
-               comment: Optional[str] = None,
-               version: Optional[str] = None,
-               path: Optional[str] = None,
-               ref: Optional[str] = None,
-               commit_hash: Optional[str] = None,
-               updated_date: Optional[str | datetime] = None,
-               git_project: Optional[str] = None) -> Rule:
+    def create(
+        self,
+        customer: str,
+        name: str,
+        resource: str,
+        description: str,
+        rule_source_id: str,
+        cloud: Optional[str] = None,
+        filters: Optional[list[dict]] = None,
+        comment: Optional[str] = None,
+        path: Optional[str] = None,
+        ref: Optional[str] = None,
+        commit_hash: Optional[str] = None,
+        updated_date: Optional[str | datetime] = None,
+        git_project: Optional[str] = None,
+    ) -> Rule:
         if isinstance(updated_date, datetime):
             updated_date = utc_iso(updated_date)
-        version = version or self.model_class.latest_version_tag()
         return super().create(
-            id=self.gen_rule_id(customer, cloud, name, version),
+            id=self.gen_rule_id(customer, cloud, name, LATEST_VERSION_TAG),
             customer=customer,
             resource=resource,
             description=description,
@@ -293,16 +324,20 @@ class RuleService(BaseDataService[Rule]):
             location=self.gen_location(git_project, ref, path),
             commit_hash=commit_hash,
             updated_date=updated_date,
-            rule_source_id=rule_source_id
+            rule_source_id=rule_source_id,
         )
 
     def get_by_id_index(
-            self, customer: str, cloud: Optional[str] = None,
-            name: Optional[str] = None, version: Optional[str] = None,
-            ascending: Optional[bool] = False, limit: Optional[int] = None,
-            last_evaluated_key: Optional[dict] = None,
-            filter_condition: Optional[Condition] = None,
-            attributes_to_get: Optional[list] = None) -> ResultIterator[Rule]:
+        self,
+        customer: str,
+        cloud: Optional[str] = None,
+        name: Optional[str] = None,
+        ascending: Optional[bool] = False,
+        limit: Optional[int] = None,
+        last_evaluated_key: Optional[dict] = None,
+        filter_condition: Optional[Condition] = None,
+        attributes_to_get: Optional[list] = None,
+    ) -> Iterator[Rule]:
         """
         Performs query by rules with full match of provided parameters.
         This query uses Customer id index (c-id-index)
@@ -311,7 +346,6 @@ class RuleService(BaseDataService[Rule]):
         :param customer:
         :param cloud:
         :param name:
-        :param version:
         :param ascending:
         :param limit:
         :param last_evaluated_key:
@@ -322,12 +356,14 @@ class RuleService(BaseDataService[Rule]):
         _LOG.info('Going to query rules')
         if name and not cloud:
             # custom behaviour for this method if cloud not given
-            _LOG.warning('Cloud was not provided but name was. '
-                         'Trying to resolve cloud from name')
+            _LOG.warning(
+                'Cloud was not provided but name was. '
+                'Trying to resolve cloud from name'
+            )
             cloud = RuleName(name).cloud.value
             if not cloud:
                 return EmptyResultIterator(last_evaluated_key)
-        sort_key = self.gen_rule_id(customer, cloud, name, version)
+        sort_key = self.gen_rule_id(customer, cloud, name)
         return self.model_class.customer_id_index.query(
             hash_key=customer,
             range_key_condition=self.model_class.id.startswith(sort_key),
@@ -335,53 +371,67 @@ class RuleService(BaseDataService[Rule]):
             limit=limit,
             last_evaluated_key=last_evaluated_key,
             filter_condition=filter_condition,
-            attributes_to_get=attributes_to_get
+            attributes_to_get=attributes_to_get,
         )
 
-    def get_by_rule_source_id(self, rule_source_id: str, customer: str,
-                              cloud: str | None = None, ascending: bool = True,
-                              limit: int | None = None,
-                              last_evaluated_key: dict | None = None,
-                              ) -> Iterator[Rule]:
+    def get_by_rule_source_id(
+        self,
+        rule_source_id: str,
+        customer: str,
+        cloud: str | None = None,
+        ascending: bool = True,
+        limit: int | None = None,
+        last_evaluated_key: dict | None = None,
+    ) -> Iterator[Rule]:
         sort_key = self.gen_rule_id(customer, cloud)
         return self.model_class.rule_source_id_id_index.query(
             hash_key=rule_source_id,
             range_key_condition=self.model_class.id.startswith(sort_key),
             scan_index_forward=ascending,
             limit=limit,
-            last_evaluated_key=last_evaluated_key
+            last_evaluated_key=last_evaluated_key,
         )
 
-    def get_by_rule_source(self, rule_source: RuleSource,
-                           cloud: str | None = None, ascending: bool = True,
-                           limit: int | None = None,
-                           last_evaluated_key: dict | None = None
-                           ) -> Iterator[Rule]:
+    def get_by_rule_source(
+        self,
+        rule_source: RuleSource,
+        cloud: str | None = None,
+        ascending: bool = True,
+        limit: int | None = None,
+        last_evaluated_key: dict | None = None,
+    ) -> Iterator[Rule]:
         return self.get_by_rule_source_id(
             rule_source_id=rule_source.id,
             customer=rule_source.customer,
             cloud=cloud,
             ascending=ascending,
             limit=limit,
-            last_evaluated_key=last_evaluated_key
+            last_evaluated_key=last_evaluated_key,
         )
 
-    def get_latest_rule(self, customer: str, name: str,
-                        cloud: Optional[str] = None,
-                        ) -> Optional[Rule]:
-        return next(self.get_by_id_index(
-            customer=customer,
-            cloud=cloud,
-            name=name,
-            ascending=False,
-            limit=1
-        ), None)
+    def get_latest_rule(
+        self, customer: str, name: str, cloud: Optional[str] = None
+    ) -> Optional[Rule]:
+        return next(
+            self.get_by_id_index(
+                customer=customer,
+                cloud=cloud,
+                name=name,
+                ascending=False,
+                limit=1,
+            ),
+            None,
+        )
 
-    def get_fuzzy_by(self, customer: str, name_prefix: str,
-                     cloud: Optional[str] = None, ascending: bool = False,
-                     limit: Optional[int] = None,
-                     last_evaluated_key: Optional[dict] = None
-                     ) -> Iterator[Rule]:
+    def get_fuzzy_by(
+        self,
+        customer: str,
+        name_prefix: str,
+        cloud: Optional[str] = None,
+        ascending: bool = False,
+        limit: Optional[int] = None,
+        last_evaluated_key: Optional[dict] = None,
+    ) -> Iterator[Rule]:
         """
         Looks for rules by the given rule name prefix (ecc, ecc-aws,
         ecc-aws-022).
@@ -394,23 +444,28 @@ class RuleService(BaseDataService[Rule]):
         :return:
         """
         if not cloud:
-            _LOG.warning('Cloud was not given to get_fuzzy_by. '
-                         'Trying to resolve from name')
+            _LOG.warning(
+                'Cloud was not given to get_fuzzy_by. '
+                'Trying to resolve from name'
+            )
             cloud = RuleName(name_prefix).cloud.value
             if not cloud:
                 return EmptyResultIterator(last_evaluated_key)
-        sort_key = f'{customer}{COMPOUND_KEYS_SEPARATOR}{cloud}' \
-                   f'{COMPOUND_KEYS_SEPARATOR}{name_prefix}'
+        sort_key = (
+            f'{customer}{COMPOUND_KEYS_SEPARATOR}{cloud}'
+            f'{COMPOUND_KEYS_SEPARATOR}{name_prefix}'
+        )
         return self.model_class.customer_id_index.query(
             hash_key=customer,
             range_key_condition=self.model_class.id.startswith(sort_key),
             scan_index_forward=ascending,
             limit=limit,
-            last_evaluated_key=last_evaluated_key
+            last_evaluated_key=last_evaluated_key,
         )
 
-    def resolve_rule(self, customer: str, name_prefix: str,
-                     cloud: Optional[str] = None) -> Optional[Rule]:
+    def resolve_rule(
+        self, customer: str, name_prefix: str, cloud: Optional[str] = None
+    ) -> Optional[Rule]:
         """
         Looks for a concrete rule by given name prefix. If found, the latest
         version of the rule is returned.
@@ -421,18 +476,21 @@ class RuleService(BaseDataService[Rule]):
         :param cloud:
         :return:
         """
-        return next(self.get_fuzzy_by(
-            customer=customer,
-            name_prefix=name_prefix,
-            cloud=cloud,
-            ascending=False,
-            limit=1
-        ), None)
+        return next(
+            self.get_fuzzy_by(
+                customer=customer,
+                name_prefix=name_prefix,
+                cloud=cloud,
+                ascending=False,
+                limit=1,
+            ),
+            None,
+        )
 
     @staticmethod
-    def without_duplicates(rules: Iterable[Rule],
-                           rules_version: Optional[str] = None
-                           ) -> Generator[Rule, None, None]:
+    def without_duplicates(
+        rules: Iterable[Rule], rules_version: Optional[str] = None
+    ) -> Generator[Rule, None, None]:
         """
         If ruleset contains different version of the same rules,
         the latest version will be kept. If rules_version is specified,
@@ -445,14 +503,19 @@ class RuleService(BaseDataService[Rule]):
                 name_rule[_name] = rule
                 continue
             # duplicate
-            if isinstance(rules_version,
-                          str) and rule.version == rules_version:
+            if (
+                isinstance(rules_version, str)
+                and rule.version == rules_version
+            ):
                 name_rule[_name] = rule
                 continue
             # duplicate and either no rules_version or
             # rule.version != rules_version
-            if (name_rule[_name].version != rules_version and
-                    name_rule[_name].version < rule.version):
+            if (
+                name_rule[_name].version != rules_version
+                and name_rule[_name].normalized_version
+                < rule.normalized_version
+            ):
                 name_rule[_name] = rule  # override with the largest version
         yield from name_rule.values()
 
@@ -471,12 +534,16 @@ class RuleService(BaseDataService[Rule]):
         return dct
 
     def get_by_location_index(
-            self, customer: str, project: Optional[str] = None,
-            ref: Optional[str] = None, path: Optional[str] = None,
-            ascending: bool = False, limit: Optional[int] = None,
-            last_evaluated_key: Optional[dict] = None,
-            filter_condition: Optional[Condition] = None
-    ) -> ResultIterator[Rule]:
+        self,
+        customer: str,
+        project: Optional[str] = None,
+        ref: Optional[str] = None,
+        path: Optional[str] = None,
+        ascending: bool = False,
+        limit: Optional[int] = None,
+        last_evaluated_key: Optional[dict] = None,
+        filter_condition: Optional[Condition] = None,
+    ) -> Iterator[Rule]:
         """
         This query uses Customer location index (c-l-index).
         This is a low-level implementation when filter can be provided
@@ -494,21 +561,26 @@ class RuleService(BaseDataService[Rule]):
         sk = self.gen_location(project, ref, path)
         return self.model_class.customer_location_index.query(
             hash_key=customer,
-            range_key_condition=(self.model_class.location.startswith(sk)),
+            range_key_condition=self.model_class.location.startswith(sk),
             scan_index_forward=ascending,
             limit=limit,
             last_evaluated_key=last_evaluated_key,
-            filter_condition=filter_condition
+            filter_condition=filter_condition,
         )
 
-    def get_by(self, customer: str, project: Optional[str] = None,
-               ref: Optional[str] = None, path: Optional[str] = None,
-               cloud: Optional[str] = None, name: Optional[str] = None,
-               version: Optional[str] = None,
-               ascending: bool = False, limit: Optional[int] = None,
-               last_evaluated_key: Optional[dict] = None,
-               index: Literal['c-l-index', 'c-id-index'] = 'c-l-index'
-               ) -> ResultIterator[Rule]:
+    def get_by(
+        self,
+        customer: str,
+        project: Optional[str] = None,
+        ref: Optional[str] = None,
+        path: Optional[str] = None,
+        cloud: Optional[str] = None,
+        name: Optional[str] = None,
+        ascending: bool = False,
+        limit: Optional[int] = None,
+        last_evaluated_key: Optional[dict] = None,
+        index: Literal['c-l-index', 'c-id-index'] = 'c-l-index',
+    ) -> Iterator[Rule]:
         """
         A hybrid between get_by_id_index and get_by_location_index.
         This method can use either index. Which one will perform more
@@ -521,7 +593,6 @@ class RuleService(BaseDataService[Rule]):
         :param path:
         :param cloud:
         :param name:
-        :param version:
         :param ascending:
         :param limit:
         :param last_evaluated_key:
@@ -531,7 +602,7 @@ class RuleService(BaseDataService[Rule]):
         assert index in {'c-l-index', 'c-id-index'}
         if index == 'c-l-index':
             _LOG.debug('Querying using c-l-index')
-            _id = self.gen_rule_id(customer, cloud, name, version)
+            _id = self.gen_rule_id(customer, cloud, name)
             return self.get_by_location_index(
                 customer=customer,
                 project=project,
@@ -540,7 +611,7 @@ class RuleService(BaseDataService[Rule]):
                 ascending=ascending,
                 limit=limit,
                 last_evaluated_key=last_evaluated_key,
-                filter_condition=self.model_class.id.startswith(_id)
+                filter_condition=self.model_class.id.startswith(_id),
             )
         # index == 'c-id-index'
         condition = None
@@ -551,43 +622,40 @@ class RuleService(BaseDataService[Rule]):
             customer=customer,
             cloud=cloud,
             name=name,
-            version=version,
             ascending=ascending,
             limit=limit,
             last_evaluated_key=last_evaluated_key,
-            filter_condition=condition
+            filter_condition=condition,
         )
 
     @staticmethod
-    def filter_by(rules: Iterable[Rule],
-                  customer: Optional[FilterValue] = None,
-                  cloud: Optional[FilterValue] = None,
-                  name_prefix: Optional[FilterValue] = None,
-                  version: Optional[FilterValue] = None,
-                  git_project: Optional[FilterValue] = None,
-                  ref: Optional[FilterValue] = None,
-                  rule_source_id: Optional[FilterValue] = None,
-                  resource: Optional[FilterValue] = None
-                  ) -> Iterator[Rule]:
+    def filter_by(
+        rules: Iterable[Rule],
+        customer: Optional[FilterValue] = None,
+        cloud: Optional[FilterValue] = None,
+        name_prefix: Optional[FilterValue] = None,
+        git_project: Optional[FilterValue] = None,
+        ref: Optional[FilterValue] = None,
+        rule_source_id: Optional[FilterValue] = None,
+        resource: Optional[FilterValue] = None,
+    ) -> Iterator[Rule]:
         """
         God-like filter. Filter just using python. No queries
         """
         if isinstance(customer, str):
-            customer = (customer, )
+            customer = (customer,)
         if isinstance(cloud, str):
-            cloud = (cloud, )
+            cloud = (cloud,)
         if isinstance(name_prefix, str):
-            name_prefix = (name_prefix, )
-        if isinstance(version, str):
-            name_prefix = (version, )
+            name_prefix = (name_prefix,)
         if isinstance(git_project, str):
-            git_project = (git_project, )
+            git_project = (git_project,)
         if isinstance(ref, str):
-            ref = (ref, )
+            ref = (ref,)
         if isinstance(resource, str):
-            resource = (resource, )
+            resource = (resource,)
         if isinstance(rule_source_id, str):
-            rule_source_id = (rule_source_id, )
+            rule_source_id = (rule_source_id,)
 
         def _check(rule: Rule) -> bool:
             if customer and rule.customer not in customer:
@@ -595,7 +663,8 @@ class RuleService(BaseDataService[Rule]):
             if cloud and rule.cloud not in map(adjust_cloud, cloud):
                 return False
             if name_prefix and not any(
-                    map(lambda n: rule.name.startswith(n), name_prefix)):
+                map(lambda n: rule.name.startswith(n), name_prefix)
+            ):
                 return False
             if git_project and rule.git_project not in git_project:
                 return False